[rpc] fix unwrap
